Extract MUI system config to a separate directory
