+++ package.json
@@ -15,7 +15,7 @@
     "release:publish:dry-run": "lerna publish from-package --dist-tag latest --contents build --registry=\"http://localhost:4873/\"",
     "release:tag": "node scripts/releaseTag.mjs",
     "docs:api": "rimraf --glob ./docs/pages/**/api-docs ./docs/pages/**/api && yarn docs:api:build",
-    "docs:api:build": "ts-node ./packages/api-docs-builder/buildApi.ts",
+    "docs:api:build": "ts-node ./packages/api-docs-builder/cli.ts",
     "docs:build": "yarn workspace docs build",
     "docs:build-sw": "yarn workspace docs build-sw",
     "docs:build-color-preview": "babel-node scripts/buildColorTypes",
